requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt ./requirements.in
#
aniso8601==9.0.1
    # via graphene
asgiref==3.8.1
    # via django
django==4.2.14
    # via
    #   -r ./requirements.in
    #   django-filter
    #   django-graphql-jwt
    #   graphene-django
django-filter==24.2
    # via -r ./requirements.in
django-graphql-jwt==0.4.0
    # via -r ./requirements.in
graphene==3.3
    # via
    #   django-graphql-jwt
    #   graphene-django
graphene-django==3.2.2
    # via django-graphql-jwt
graphql-core==3.2.3
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
promise==2.3
    # via graphene-django
psycopg2==2.9.9
    # via -r ./requirements.in
pycryptodome==3.20.0
    # via -r ./requirements.in
pyjwt==2.8.0
    # via django-graphql-jwt
six==1.16.0
    # via promise
sqlparse==0.5.0
    # via django
text-unidecode==1.3
    # via graphene-django
